// vue/src/store/generated/defund-labs/defund/defundlabs.defund.interquery/module/rest.ts
export interface InterqueryInterquery {
  creator?: string;
  storeid?: string;
  path?: string;

  /** @format byte */
  key?: string;

  /** @format uint64 */
  timeoutHeight?: string;
  clientId?: string;
}

export interface InterqueryInterqueryResult {
  creator?: string;
  storeid?: string;

  /** @format byte */
  data?: string;

  /** @format uint64 */
  height?: string;
  clientId?: string;
  success?: boolean;
  proof?: CryptoProofOps;
}

export interface InterqueryInterqueryTimeoutResult {
  creator?: string;
  storeid?: string;

  /** @format uint64 */
  timeoutHeight?: string;
  clientId?: string;
}

export type InterqueryMsgCreateInterqueryResponse = object;

export type InterqueryMsgCreateInterqueryResultResponse = object;

export type InterqueryMsgCreateInterqueryTimeoutResponse = object;

export interface InterqueryQueryAllInterqueryResponse {
  interquery?: InterqueryInterquery[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface InterqueryQueryAllInterqueryResultResponse {
  interqueryresult?: InterqueryInterqueryResult[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface InterqueryQueryAllInterqueryTimeoutResultResponse {
  interquerytimeoutresult?: InterqueryInterqueryTimeoutResult[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;
}

export interface InterqueryQueryGetInterqueryResponse {
  interquery?: InterqueryInterquery;
}

export interface InterqueryQueryGetInterqueryResultResponse {
  interqueryresult?: InterqueryInterqueryResult;
}

export interface InterqueryQueryGetInterqueryTimeoutResultResponse {
  interquerytimeoutresult?: InterqueryInterqueryTimeoutResult;
}

export interface ProtobufAny {
  "@type"?: string;
}

export interface RpcStatus {
  /** @format int32 */
  code?: number;
  message?: string;
  details?: ProtobufAny[];
}
// ...
export interface V1Beta1PageResponse {
  /** @format byte */
  next_key?: string;

  /** @format uint64 */
  total?: string;
}

export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;

export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: keyof Omit<Body, "body" | "bodyUsed">;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;

export interface ApiConfig<SecurityDataType = unknown> {
  baseUrl?: string;
  baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">;
  securityWorker?: (securityData: SecurityDataType) => RequestParams | void;
}

export interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
  data: D;
  error: E;
}

type CancelToken = Symbol | string | number;

export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
}

export class HttpClient<SecurityDataType = unknown> {
  public baseUrl: string = "";
  private securityData: SecurityDataType = null as any;
  private securityWorker: null | ApiConfig<SecurityDataType>["securityWorker"] = null;
  private abortControllers = new Map<CancelToken, AbortController>();

  private baseApiParams: RequestParams = {
    credentials: "same-origin",
    headers: {},
    redirect: "follow",
    referrerPolicy: "no-referrer",
  };

  constructor(apiConfig: ApiConfig<SecurityDataType> = {}) {
    Object.assign(this, apiConfig);
  }

  public setSecurityData = (data: SecurityDataType) => {
    this.securityData = data;
  };

  private addQueryParam(query: QueryParamsType, key: string) {
    const value = query[key];

    return (
      encodeURIComponent(key) +
      "=" +
      encodeURIComponent(Array.isArray(value) ? value.join(",") : typeof value === "number" ? value : `${value}`)
    );
  }

  protected toQueryString(rawQuery?: QueryParamsType): string {
    const query = rawQuery || {};
    const keys = Object.keys(query).filter((key) => "undefined" !== typeof query[key]);
    return keys
      .map((key) =>
        typeof query[key] === "object" && !Array.isArray(query[key])
          ? this.toQueryString(query[key] as QueryParamsType)
          : this.addQueryParam(query, key),
      )
      .join("&");
  }

  protected addQueryParams(rawQuery?: QueryParamsType): string {
    const queryString = this.toQueryString(rawQuery);
    return queryString ? `?${queryString}` : "";
  }

  private contentFormatters: Record<ContentType, (input: any) => any> = {
    [ContentType.Json]: (input: any) =>
      input !== null && (typeof input === "object" || typeof input === "string") ? JSON.stringify(input) : input,
    [ContentType.FormData]: (input: any) =>
      Object.keys(input || {}).reduce((data, key) => {
        data.append(key, input[key]);
        return data;
      }, new FormData()),
    [ContentType.UrlEncoded]: (input: any) => this.toQueryString(input),
  };

  private mergeRequestParams(params1: RequestParams, params2?: RequestParams): RequestParams {
    return {
      ...this.baseApiParams,
      ...params1,
      ...(params2 || {}),
      headers: {
        ...(this.baseApiParams.headers || {}),
        ...(params1.headers || {}),
        ...((params2 && params2.headers) || {}),
      },
    };
  }

  private createAbortSignal = (cancelToken: CancelToken): AbortSignal | undefined => {
    if (this.abortControllers.has(cancelToken)) {
      const abortController = this.abortControllers.get(cancelToken);
      if (abortController) {
        return abortController.signal;
      }
      return void 0;
    }

    const abortController = new AbortController();
    this.abortControllers.set(cancelToken, abortController);
    return abortController.signal;
  };

  public abortRequest = (cancelToken: CancelToken) => {
    const abortController = this.abortControllers.get(cancelToken);

    if (abortController) {
      abortController.abort();
      this.abortControllers.delete(cancelToken);
    }
  };

  public request = <T = any, E = any>({
    body,
    secure,
    path,
    type,
    query,
    format = "json",
    baseUrl,
    cancelToken,
    ...params
  }: FullRequestParams): Promise<HttpResponse<T, E>> => {
    const secureParams = (secure && this.securityWorker && this.securityWorker(this.securityData)) || {};
    const requestParams = this.mergeRequestParams(params, secureParams);
    const queryString = query && this.toQueryString(query);
    const payloadFormatter = this.contentFormatters[type || ContentType.Json];

    return fetch(`${baseUrl || this.baseUrl || ""}${path}${queryString ? `?${queryString}` : ""}`, {
      ...requestParams,
      headers: {
        ...(type && type !== ContentType.FormData ? { "Content-Type": type } : {}),
        ...(requestParams.headers || {}),
      },
      signal: cancelToken ? this.createAbortSignal(cancelToken) : void 0,
      body: typeof body === "undefined" || body === null ? null : payloadFormatter(body),
    }).then(async (response) => {
      const r = response as HttpResponse<T, E>;
      r.data = (null as unknown) as T;
      r.error = (null as unknown) as E;

      const data = await response[format]()
        .then((data) => {
          if (r.ok) {
            r.data = data;
          } else {
            r.error = data;
          }
          return r;
        })
        .catch((e) => {
          r.error = e;
          return r;
        });

      if (cancelToken) {
        this.abortControllers.delete(cancelToken);
      }

      if (!response.ok) throw data;
      return data;
    });
  };
}

/**
 * @title interquery/genesis.proto
 * @version version not set
 */
export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
  /**
   * No description
   *
   * @tags Query
   * @name QueryInterqueryAll
   * @summary Queries a list of interquery items.
   * @request GET:/defund-labs/defund/interquery/interquery
   */
  queryInterqueryAll = (
    query?: {
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.count_total"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<InterqueryQueryAllInterqueryResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interquery`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryInterquery
   * @summary Queries a interquery by index.
   * @request GET:/defund-labs/defund/interquery/interquery/{storeid}
   */
  queryInterquery = (storeid: string, params: RequestParams = {}) =>
    this.request<InterqueryQueryGetInterqueryResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interquery/${storeid}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryInterqueryResultAll
   * @summary Queries a list of interquery result items.
   * @request GET:/defund-labs/defund/interquery/interqueryresult
   */
  queryInterqueryResultAll = (
    query?: {
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.count_total"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<InterqueryQueryAllInterqueryResultResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interqueryresult`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryInterqueryResult
   * @summary Queries a interquery result by index.
   * @request GET:/defund-labs/defund/interquery/interqueryresult/{storeid}
   */
  queryInterqueryResult = (storeid: string, params: RequestParams = {}) =>
    this.request<InterqueryQueryGetInterqueryResultResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interqueryresult/${storeid}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryInterqueryTimeoutResultAll
   * @summary Queries a list of interquery timeout result items.
   * @request GET:/defund-labs/defund/interquery/interquerytimeoutresult
   */
  queryInterqueryTimeoutResultAll = (
    query?: {
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.count_total"?: boolean;
      "pagination.reverse"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<InterqueryQueryAllInterqueryTimeoutResultResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interquerytimeoutresult`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryInterqueryTimeoutResult
   * @summary Queries a interquery timeout result by index.
   * @request GET:/defund-labs/defund/interquery/interquerytimeoutresult/{storeid}
   */
  queryInterqueryTimeoutResult = (storeid: string, params: RequestParams = {}) =>
    this.request<InterqueryQueryGetInterqueryTimeoutResultResponse, RpcStatus>({
      path: `/defund-labs/defund/interquery/interquerytimeoutresult/${storeid}`,
      method: "GET",
      format: "json",
      ...params,
    });
}